+++ changelog.md
@@ -3,6 +3,8 @@
 - Rotation Support
 - Fixes: [#268](../../issues/268), [#284](../../issues/284)
 - Adding support for disabling native events for sendKeys command to support e.g. German umlauts (#7,#110)
+- Added support for sending key events, send text and tap via adb connection
+
 
 0.8.0
 ---